+++ sdk/src/main/java/com/ingresse/sdk/IngresseService.kt
@@ -11,6 +11,7 @@ class IngresseService(client: IngresseClient) {
     var ticketList = TicketListService(client)
     var ticketStatus = TicketStatusService(client)
     var transaction = TransactionService(client)
+    var transfer = TransferService(client)
     var user = UserService(client)
     var event = EventService(client)
     var ticket = TicketService(client)
