+++ sdk/src/main/java/com/ingresse/sdk/model/response/FriendsFromSearchJSON.kt
@@ -2,7 +2,7 @@ package com.ingresse.sdk.model.response
 
 data class FriendsFromSearchJSON(
     val email: String? = "",
-    val id: Int? = 0,
+    val id: Long? = 0,
     val name: String? = "",
     val picture: String? = ""
 )
