+++ sdk/src/main/java/com/ingresse/sdk/request/Search.kt
@@ -20,18 +20,4 @@ interface Search {
                         @Query("from") from: String,
                         @Query("orderBy") orderBy: String,
                         @Query("offset") offset: String) : Call<String>
-
-    /**
-     * Search friends by name
-     *
-     * @param term - String term to search for
-     * @param size - Number of results
-     * @param userToken - User token
-     */
-    @GET("/search/transfer/user")
-    fun getFriendsFromSearch(@Query("term") term: String,
-                             @Query("size") size: String?,
-                             @Query("apikey") apikey: String,
-                             @Query("usertoken") userToken: String): Call<String>
-
 }
