+++ sdk/src/main/java/com/ingresse/sdk/request/User.kt
@@ -45,38 +45,4 @@ interface User {
                        @Query("page") page: Int? = null,
                        @Query("pageSize") pageSize: Int? = null,
                        @Query("usertoken") token: String): Call<String>
-
-    /**
-     * Gets user transfers.
-     *
-     * @param userId id from user
-     * @param page - page from user ticket
-     * @param pageSize - size of user ticket page
-     * @param usertoken - token from user
-     * @param status - user transfers status
-     */
-    @GET("/user/{userId}/transfers")
-    fun getUserTransfers(@Path("userId") userId: String,
-                         @Query("apikey") apikey: String,
-                         @Query("page") page: Int? = null,
-                         @Query("pageSize") pageSize: Int? = null,
-                         @Query("usertoken") token: String,
-                         @Query("status") status: String): Call<String>
-
-
-    /**
-     * Gets user recent transfers.
-     *
-     * @param userId id of the user.
-     * @param userToken - User token
-     * @param order - order control for the result
-     * @param size - result page size
-     */
-    @GET("/user/{id}/last-transfers")
-    fun getRecentTransfers(@Path("userId") userId: String,
-                           @Query("apikey") apikey: String,
-                           @Query("usertoken") userToken: String,
-                           @Query("order") order: String?,
-                           @Query("size") size: Int?): Call<String>
-
 }
