+++ sdk/src/main/java/com/ingresse/sdk/services/SearchService.kt
@@ -23,8 +23,6 @@ class SearchService(private val client: IngresseClient) {
     private var cancelAllCalled = false
 
     private var mEventSearchCall: Call<String>? = null
-    private var mFriendsFromSearchCall: Call<String>? = null
-    private var mConcurrentCalls: ArrayList<Call<String>> = ArrayList()
 
     init {
         val httpClient = ClientBuilder(client)
@@ -41,33 +39,9 @@ class SearchService(private val client: IngresseClient) {
     }
 
     /**
-     * Method to cancel all requests
-     */
-    fun cancelAll() {
-        cancelAllCalled = true
-        mEventSearchCall?.cancel()
-        mFriendsFromSearchCall?.cancel()
-        mConcurrentCalls.forEach { it.cancel() }
-        mConcurrentCalls.clear()
-    }
-
-    /**
      * Method to cancel a event search request
      */
     fun cancelEventSearchByTitle() = mEventSearchCall?.cancel()
-
-    /**
-     * Method to cancel a get friends from search
-     */
-    fun cancelFriendsFromSearch(concurrent: Boolean = false) {
-        if (!concurrent) {
-            mFriendsFromSearchCall?.cancel()
-            return
-        }
-
-        mConcurrentCalls.forEach { it.cancel() }
-        mConcurrentCalls.clear()
-    }
 
     /**
      * Search event by title
@@ -104,56 +78,5 @@ class SearchService(private val client: IngresseClient) {
 
         val type = object: TypeToken<ResponseHits<EventJSON>?>() {}.type
         mEventSearchCall?.enqueue(RetrofitCallback(type, callback))
-    }
-
-    /**
-     * Search friends by name
-     *
-     * @param concurrent - parameters to concurrent request
-     * @param request - parameters required to request
-     * @param onSuccess - success callback
-     * @param onError - error callback
-     * @param onConnectionError - connection error callback
-     */
-    fun getFriendsFromSearch(concurrent: Boolean = false,
-                             request: FriendsFromSearch,
-                             onSuccess: (Array<FriendsFromSearchJSON>) -> Unit,
-                             onError: (APIError) -> Unit,
-                             onConnectionError: (error: Throwable) -> Unit) {
-
-        var call = service.getFriendsFromSearch(
-            term = request.term,
-            size = request.size,
-            apikey = client.key,
-            userToken = request.usertoken
-        )
-
-        if (!concurrent) mFriendsFromSearchCall = call else mConcurrentCalls.add(call)
-
-        val callback = object: IngresseCallback<Response<Array<FriendsFromSearchJSON>>?> {
-            override fun onSuccess(data: Response<Array<FriendsFromSearchJSON>>?) {
-                val response = data?.responseData ?: return onError(APIError.default)
-
-                if (!concurrent) mFriendsFromSearchCall = null else mConcurrentCalls.remove(call)
-                onSuccess(response)
-            }
-
-            override fun onError(error: APIError) {
-                if (!concurrent) mFriendsFromSearchCall = null else mConcurrentCalls.remove(call)
-                onError(error)
-            }
-
-            override fun onRetrofitError(error: Throwable) {
-                if (!concurrent) mFriendsFromSearchCall = null else mConcurrentCalls.remove(call)
-                if (error is IOException) return onConnectionError(error)
-
-                val apiError = APIError()
-                apiError.message = error.localizedMessage
-                onError(apiError)
-            }
-        }
-
-        val type = object : TypeToken<Response<Array<FriendsFromSearchJSON>>?>() {}.type
-        call.enqueue(RetrofitCallback(type, callback))
     }
 }
