+++ sdk/src/main/java/com/ingresse/sdk/services/UserService.kt
@@ -22,8 +22,6 @@ class UserService(private val client: IngresseClient) {
     private var mUserDataCall: Call<String>? = null
     private var mUpdateBasicInfosCall: Call<String>? = null
     private var mUserTicketsCall: Call<String>? = null
-    private var mUserTransfersCall: Call<String>? = null
-    private var mRecentTransfersCall: Call<String>? = null
     private var mConcurrentCalls: ArrayList<Call<String>> = ArrayList()
 
     init {
@@ -57,29 +55,6 @@ class UserService(private val client: IngresseClient) {
     fun cancelUserTicketsData(concurrent: Boolean = false) {
         if(!concurrent) {
             mUserTicketsCall?.cancel()
-            return
-        }
-
-        mConcurrentCalls.forEach { it.cancel() }
-        mConcurrentCalls.clear()
-    }
-
-    /**
-     * Method to cancel transfers data request
-     */
-    fun cancelUserTransfersData(concurrent: Boolean = false) {
-        if(!concurrent) {
-            mUserTransfersCall?.cancel()
-            return
-        }
-
-        mConcurrentCalls.forEach { it.cancel() }
-        mConcurrentCalls.clear()
-    }
-
-    fun cancelRecentTransfersData(concurrent: Boolean = false) {
-        if(!concurrent) {
-            mRecentTransfersCall?.cancel()
             return
         }
 
@@ -232,111 +207,6 @@ class UserService(private val client: IngresseClient) {
         }
 
         val type = object : TypeToken<Response<Array<UserTicketsJSON>>?>() {}.type
-        call.enqueue(RetrofitCallback(type, callback))
-    }
-
-    /**
-     * Get user transfers data
-     *
-     * @param concurrent - parameters to concurrent request
-     * @param request - parameters required to request
-     * @param onSuccess - success callback
-     * @param onError - error callback
-     * @param onConnectionError - connection error callback
-     */
-    fun getUserTransfersData(concurrent: Boolean = false,
-                             request: UserTransfersData,
-                             onSuccess: (Array<UserTransfersJSON>) -> Unit,
-                             onError: (APIError) -> Unit,
-                             onConnectionError: (error: Throwable) -> Unit) {
-
-        var call = service.getUserTransfers(
-            userId = request.userId,
-            apikey = client.key,
-            page = request.page,
-            pageSize = request.pageSize,
-            token = request.usertoken,
-            status = request.status
-        )
-
-        if (!concurrent) mUserTransfersCall = call else mConcurrentCalls.add(call)
-
-        val callback = object: IngresseCallback<Response<Array<UserTransfersJSON>>?> {
-            override fun onSuccess(data: Response<Array<UserTransfersJSON>>?) {
-                val response = data?.responseData ?: return onError(APIError.default)
-
-                if(!concurrent) mUserTransfersCall = null else mConcurrentCalls.remove(call)
-                onSuccess(response)
-            }
-
-            override fun onError(error: APIError) {
-                if (!concurrent) mUserTransfersCall = null else mConcurrentCalls.remove(call)
-                onError(error)
-            }
-
-            override fun onRetrofitError(error: Throwable) {
-                if (!concurrent) mUserTransfersCall = null else mConcurrentCalls.remove(call)
-                if (error is IOException) return onConnectionError(error)
-
-                val apiError = APIError()
-                apiError.message = error.localizedMessage
-                onError(apiError)
-            }
-        }
-
-        val type = object : TypeToken<Response<Array<UserTransfersJSON>>?>() {}.type
-        call.enqueue(RetrofitCallback(type, callback))
-    }
-
-    /**
-     * Gets user recent transfers data
-     *
-     * @param concurrent - parameters to concurrent request
-     * @param request - parameters required to request
-     * @param onSuccess - success callback
-     * @param onError - error callback
-     * @param onConnectionError - connection error callback
-     */
-    fun getRecentTransfersData(concurrent: Boolean = false,
-                               request: RecentTransfers,
-                               onSuccess: (List<RecentTransfersJSON>) -> Unit,
-                               onError: (APIError) -> Unit,
-                               onConnectionError: (error: Throwable) -> Unit) {
-
-        var call = service.getRecentTransfers(
-            userId = request.userId,
-            apikey = client.key,
-            userToken = request.usertoken,
-            order = request.order,
-            size = request.size
-        )
-
-        if (!concurrent) mRecentTransfersCall = call else mConcurrentCalls.add(call)
-
-        val callback = object: IngresseCallback<Response<List<RecentTransfersJSON>>?> {
-            override fun onSuccess(data: Response<List<RecentTransfersJSON>>?) {
-                val response = data?.responseData ?: return onError(APIError.default)
-
-                if(!concurrent) mRecentTransfersCall = null else mConcurrentCalls.remove(call)
-                onSuccess(response)
-            }
-
-            override fun onError(error: APIError) {
-                if (!concurrent) mRecentTransfersCall = null else mConcurrentCalls.remove(call)
-                onError(error)
-            }
-
-            override fun onRetrofitError(error: Throwable) {
-                if (!concurrent) mRecentTransfersCall = null else mConcurrentCalls.remove(call)
-                if (error is IOException) return onConnectionError(error)
-
-                val apiError = APIError()
-                apiError.message = error.localizedMessage
-                onError(apiError)
-            }
-        }
-
-        val type = object : TypeToken<Response<List<RecentTransfersJSON>>?>() {}.type
         call.enqueue(RetrofitCallback(type, callback))
     }
 }
